requirements.txt
fastapi==0.95.0
pydantic==1.10.0
uvicorn==0.18.2
httpx==0.23.0
loguru==0.5.3
slowapi==0.1.7
fastapi-cache==0.4.6
redis==4.3.3
python-dotenv==0.20.0
tenacity==8.0.1
python-jose==3.3.0